Added new fields in GTFS comparison reports. Fields includes details about cracked GTFS data comparison. Updated mail service with Spring modules for better performance.

+++ projectTP/src/com/nimbler/tp/gtfs/GtfsDataMonitor.java
@@ -4,15 +4,23 @@
 package com.nimbler.tp.gtfs;
 
 import java.io.File;
+import java.io.IOException;
 import java.net.SocketException;
 import java.net.UnknownHostException;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.zip.ZipEntry;
 import java.util.zip.ZipException;
+import java.util.zip.ZipFile;
 
 import org.apache.commons.io.FilenameUtils;
+import org.apache.commons.io.IOUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.nimbler.tp.service.LoggingService;
@@ -29,8 +37,9 @@ import com.nimbler.tp.util.TpException;
  * @author nirmal
  *
  */
+@SuppressWarnings("unchecked")
 public class GtfsDataMonitor {
-
+	private static SimpleDateFormat dateFormat = new SimpleDateFormat("yyyyMMdd");
 	@Autowired
 	private LoggingService logger;
 	@Autowired
@@ -38,6 +47,8 @@ public class GtfsDataMonitor {
 	private String loggerName;
 	private String downloadDirectory;
 	List<GtfsBundle> gtfsBundles;
+	// used for debug only, must true on production
+	private static final boolean enableNewDownload = true;
 
 	public GtfsDataMonitor() {
 	}
@@ -62,7 +73,7 @@ public class GtfsDataMonitor {
 			}
 			List<String> lstAttachement = HtmlUtil.getResultTable(lstMonitorResults);
 			String strSummery = HtmlUtil.getResultSummeryTable(lstMonitorResults);
-			mailService.sendMail(TpConstants.FEEDBACK_EMAIL_ID,TpConstants.GTFS_DATA_COMPARE_MAIL_SUBJECT,strSummery,true,lstAttachement,lstAttachement);
+			mailService.sendMail(TpConstants.GTFS_DATA_COMPARE_MAIL_ID,TpConstants.GTFS_DATA_COMPARE_MAIL_SUBJECT,strSummery,true,lstAttachement,lstAttachement);
 		}catch (TpException e) {
 			logger.info(loggerName, e.getErrMsg());
 		}catch (Exception e) {
@@ -77,6 +88,7 @@ public class GtfsDataMonitor {
 	 * @return 
 	 * @throws TpException 
 	 */
+	@SuppressWarnings("unused")
 	private GtfsMonitorResult checkSingleGtfs(GtfsBundle gtfsBundle){
 		GtfsMonitorResult result = new GtfsMonitorResult(gtfsBundle);
 		String agency = gtfsBundle.getDefaultAgencyId();
@@ -85,19 +97,26 @@ public class GtfsDataMonitor {
 			logger.debug(loggerName, "Check start for "+agency+", current file:"+gtfsBundle.getCurrentDataFile());
 			String fileName = FilenameUtils.getName(currentFile);
 			File downloadFile = new File(downloadDirectory+fileName);
-			if(downloadFile.exists()){
+			if(downloadFile.exists() && enableNewDownload){
 				boolean res = downloadFile.delete();
 				logger.debug(loggerName, "Old file "+downloadFile.getAbsolutePath()+" delete: "+res);
 			}
 			if(!new File(currentFile).exists())
 				throw new TpException("No currunt GTFS file found for "+agency+" to compare");
-			logger.info(loggerName, "Downloading: "+agency+", url:"+gtfsBundle.getDownloadUrl());
-			ComUtils.getDownloadFile(downloadFile,gtfsBundle.getDownloadUrl());
-			logger.debug(loggerName, "Download complete");
-			Map<String, Date> newMap = GtfsUtils.getExpireDateFromGtfs(downloadFile);
-			Map<String, Date> oldMap = GtfsUtils.getExpireDateFromGtfs(currentFile);
+			if(enableNewDownload){
+				logger.info(loggerName, "Downloading: "+agency+", url:"+gtfsBundle.getDownloadUrl());
+				ComUtils.getDownloadFile(downloadFile,gtfsBundle.getDownloadUrl());
+				logger.debug(loggerName, "Download complete");
+			}else
+				logger.info(loggerName, "Skipping download: "+agency+", url:"+gtfsBundle.getDownloadUrl());
+			Map<String, Date> newMap = getExpireDateFromGtfs(downloadFile);
+			Map<String, Date> oldMap = getExpireDateFromGtfs(new File(currentFile));
+			Map<String, Date> crackedMap = null;
+			if(!ComUtils.isEmptyString(gtfsBundle.getCrackedDataFile()))
+				crackedMap = getExpireDateFromGtfs(new File(gtfsBundle.getCrackedDataFile()));			
 			result.setOldData(oldMap);
 			result.setNewData(newMap);
+			result.setCrackedData(crackedMap);
 			logger.info(loggerName, "File check done for "+agency+", Old service count: "+oldMap.size()+",new service count: "+newMap.size());
 		}catch (TpException e) {
 			result.setError("Error while checking gtfs file from "+agency+" : "+e.getMessage());
@@ -125,6 +144,51 @@ public class GtfsDataMonitor {
 
 	}
 	class GtfsMonitorTask{
+	}
+
+	/**
+	 * Gets the expire date from gtfs.
+	 *
+	 * @param file the file
+	 * @return the expire date from gtfs
+	 * @throws IOException Signals that an I/O exception has occurred.
+	 * @throws ParseException the parse exception
+	 * @throws TpException the tp exception
+	 */
+	public Map<String, Date> getExpireDateFromGtfs(File file) throws IOException, ParseException, TpException {
+		ZipFile zipFile = new ZipFile(file);
+		ZipEntry zipEntry = zipFile.getEntry(TpConstants.ZIP_CALENDAR_FILE);		
+		List<String> lstLines = IOUtils.readLines(zipFile.getInputStream(zipEntry),"ISO-8859-1");
+		Map<String, Date> mapServiceIdAndDate = new HashMap<String, Date>();
+		if(ComUtils.isEmptyList(lstLines))
+			throw new TpException("Invalid file, No data found in file: "+file);
+		String[] headers = lstLines.get(0).split(",");
+		int END_DATE_INDEX = -1;
+		int SERVICE_ID_INDEX = -1;
+		for (int i = 0; i < headers.length; i++) {
+			if(headers[i].toLowerCase().indexOf("service_id")!=-1)
+				SERVICE_ID_INDEX = i;
+			else if (headers[i].toLowerCase().indexOf("end_date")!=-1)
+				END_DATE_INDEX =i;
+		}
+		if(END_DATE_INDEX==-1 || SERVICE_ID_INDEX==-1)
+			throw new TpException("No service_id or end_date found in data for file:"+file);
+		for (int i = 1; i < lstLines.size(); i++) {					
+			try {
+				String[] line = lstLines.get(i).split(",");
+				if(line==null || SERVICE_ID_INDEX > (line.length-1) || END_DATE_INDEX > (line.length-1)){
+					logger.warn(loggerName, "empty line found in gtfs");
+					continue;					
+				}
+				String key =line[SERVICE_ID_INDEX];
+				String val = line[END_DATE_INDEX];
+				if(ComUtils.isEmptyString(key) || ComUtils.isEmptyString(val));
+				mapServiceIdAndDate.put(key, dateFormat.parse(StringUtils.remove(val, "\"")));
+			} catch (Exception e) {
+				logger.error(loggerName,"Malformed data Found at line "+i+", data: "+lstLines.get(i));				
+			}
+		}
+		return mapServiceIdAndDate;
 	}
 	public String getDownloadDirectory() {
 		return downloadDirectory;
